Reorganiza estrutura: separa pastas server e worker, move database para worker
