--- .gitignore
@@ -1,2 +1,0 @@
-# Ignorar a pasta database (arquivos grandes)
-database/
