+++ server/Procfile
@@ -1,0 +1,1 @@
+web: gunicorn server:app
