+++ server/resultados.txt
@@ -1,0 +1,102 @@
+hex private key: E5277388F23FC825CAF8A16801EE0CE3F0347F1E17D49E7EA2D39EDD4D8E7AAE
+WIF private key: 5KZD1rkkBzwDcMJ9UTwWQcw1wrkAWbqMkdkQsmxZkDfiV4tfJ5R
+public key: 04351D801D4FE086C9FB3BB7601BCCBA1B6EED54D2DFC53E0024E480185F711DCCBE73329FE212CB5446E638D23E787C299C6A00E2BC4642835796CAF8C89C8354
+uncompressed address: 1L5cCDfaXuDQV5Tfd3Kc7N7hxY3gntymu3
+Almost there!
+
+hex private key: 1E9813585EBBE2005710B6C4170922B1B9D32C5203C9F57AF0704A18874F606A
+WIF private key: 5J3m4k2hWF6K9jskkjZHcUKKm4KjjoGCXeqDGk1x4BCpNDQNph1
+public key: 044921EFED5164332E8CE34DD15017737563178F54DAA99A5A69B3251C402246FFEF35C4C579757FFECE2F8BFFACE1561E76F6E9F9E37B1EE4AD253B1C58E9EBAA
+uncompressed address: 1EdUx9T8S1A24S9KHgt8bJimrqWkcv577H
+Almost there!
+
+hex private key: FA7796EF265E4B5A86BFFB2F0D1E6B5FA86C363598FFC55758592FAB3B10EB30
+WIF private key: 5KibRp75BmUt3JQg9nzpdtmyeK2wDmxL78LVvDHX8CEw1NK2F8K
+public key: 04B68A72C51414F730F840F80BEBBEBE432014127062B8C197B796B2684106B25A554C5431BB0310EBC48C320021F1A50E19AF773C1EC4363C9A84CDA1AA16D61B
+uncompressed address: 1Lwip4H8wpsAogFqQEJNELo34R7Asc4yBF
+Almost there!
+
+hex private key: CF3C2E46BE9302E539C42A5F4646DE32B8D28348E956F622824D1D2C8C3E8AB7
+WIF private key: 5KPZ87mRHatzouVkErEH53qaar1cDx3mZkGM9D8Y7ssxbX5JAyQ
+public key: 04E2D0CE558FE8CEBFCECD54362760C94EA5FC4820C9FCB7088AB80C421714F2F850610BB6CC6F0E84196B2D3208A95EFAD608F6B420851CF65872889029C9DD67
+uncompressed address: 19o7AV87aFapxjX9YoUVErFJR3ZQAvgoA6
+Almost there!
+
+hex private key: 58E10A42D55FEE53D596C2E6F81C664ADFDCD291BE3955C3D197D30B64EC1DF8
+WIF private key: 5JVRssSiUP2Xv95tC6SDwrsGGLZCTwzomJX1ctUbeY9kjcWMEH2
+public key: 0410C8A7DC1862118ACF71E8CEE6E63EEC78EBF7E68536A0AB82C696FD40BB882396BCF45A65038E5162672B293F368C9A3D9627ADDEAAD2C0FC37E3C6B2D4D804
+uncompressed address: 15pVd7vo57LTdMndb1EaqqEYJkvgiiMM51
+Almost there!
+
+hex private key: 1117E541D00DC3DF570A08D55131C3766A70605C0DDBDBE307CBD73829AD64DA
+WIF private key: 5HwpD1vURFAnXSWf2kJVr6sHQPCngdYk3LgsXqe9GYZcDG6N5cU
+public key: 042347E078B34E50ABAF159CB8B363EA0C7B6B3F27C48325D8EE76F4167091BCF25AB227E4F4B274D7E38BDB986F0CDA8AF4C42171FF6E9BF7548EC85DB8C627F4
+uncompressed address: 1Q4TgCCWP7QfVWQQa8shAgjpvymQ9VsW5b
+Almost there!
+
+hex private key: A10986E625B312DE6CF2E62A1003EE27DAD2873672AD0958C152F95DF508756B
+WIF private key: 5K3D4VXteNGhxNy1SsYuMch5tQLp7KKoqnvFJ9C6eLsbVqxysXY
+public key: 04D6597D386452EFFD092AAB55BE1D1690DDCE01AB27C2423001A3BE03EEC4E45FFF9A4E7440091D3B1815A39A0307A4A98A6628ADF21F53EFF04218F6F8170084
+uncompressed address: 1PTzRZY9arCSkJeYZPxYzjBiWr7nXVSGxf
+Almost there!
+
+hex private key: 08763C278EDE8BC1C388E1E89960693BD171BA54660346C48109877BCE996348
+WIF private key: 5Ht1jC5bAJrGqJtQtTZC3R1u6TLhy1WWi5Av6Z8WcMKaTz1q6Nt
+public key: 04E77B6EFBBAA875E2A2ED40A83FB93F19575BCA6DCD0701DDB8E079DE956648996E14090442E2A2634C98DD258D407F12A4E744B416DD22FF0A7905E47D3C18E8
+uncompressed address: 14z19QWyJcwF6tBBka8ZNoufbiFWALk5CN
+Almost there!
+
+hex private key: 63B9E7D01A3CA05720EA4B16D3B1493AFDC18A74C954551F929A06ADE394EAE4
+WIF private key: 5JaCxF4pKU14DTj3dddZ4oMRNCh6QivpQDTy5EPuiPxqYJ8QuW3
+public key: 046A7586211D507F4F01DCAA304CD10FD4B47D76475FC35BED23EF5CB978A60803724DB1AF83C445E384294E28D66A2DA5896FF9F90F4714CA739D19E7F162B312
+uncompressed address: 1E9twSM5EQBqg1EFhT855McSmiRa3cKVQc
+Almost there!
+
+hex private key: AD52E946AD5B7A38819B8541762EF8A99B6FFB75F739F890241777DFB0064907
+WIF private key: 5K8cua4C8P7LxCeQosmPR6uU4Qvsy2eJn6haU1Fa8wu93daWnLy
+public key: 04F95FE92E28973CB627648C5005F8340FED114771F69E03A1E6C688392027BFA4EC852E3B6AFDC5A0E1636B767D3697C0F891A6A5ACA484DA20A09B3CFD6FE325
+uncompressed address: 12w2UpCAaKasJ9Ui5TxvgLAcsX1MKmtGfz
+Almost there!
+
+hex private key: DD5D6CBC7D4FD120F4A5E13EF786FA74276BA962CDBC82C819CE06DBD55D1C42
+WIF private key: 5KVn3xYJanb6Zw6xCEQ6jCzTS2CVVvxmEXigCfUSraSrLuxMXm6
+public key: 048E537A33DACBC4EF9DABCA900A61012F54545C191D18514CB9FC108CAFA0CB7023A4EA4942659850F2B00ECC5A1B7155E18DFA3D243774D13A4FC25280279882
+uncompressed address: 1F3Dy3ugD6SzNNfE15n3xaW1EcXcmjbjR2
+Almost there!
+
+hex private key: 2CBEE6EF29E397A731A64A42946ADBC4C3C1D8C81152971DBBE2A8668C8B126D
+WIF private key: 5J9zYtdpzcfHMXHAHiEtz4vNyY6jfSrnFHr1BvYjMdRbxH9xAsL
+public key: 0445F6F2B0E0FCE086E003EDA2984C9761AA612031DB269EC6E517AFD434BB24877C74DC7D09B07E683AEABC2B9CE38816EAEAEE7C24466DDE6171B9FB5821ABC4
+uncompressed address: 1LpUWBqXemMyAyaWXUYdive9ckmNBuw5RM
+Almost there!
+
+hex private key: 5FF8FD2E86353ECF54EBB57B1C1E15DA878832A39C4A1973752492F5F59DDEF8
+WIF private key: 5JYZ5CBU6AP1Y5tLboTrgXsy2Sy7bwi5wCmz1z4TPsH3xeKB9iD
+public key: 04146896C53046747CFE0C773AB58BA0618A56E87567D6AB7998D2CFCBC45AF1B3D81515985667B8B70B2B076F032425BBA4BE6212C23614DD6CF5C5F8751A489E
+uncompressed address: 19coUX3ZUb8sy9FwFL4NwJk5AQtqRL6KEz
+Almost there!
+
+hex private key: 73DBDEA890F5D1962A27BAE084CFD01C87464560E7CB8DFA89F690D555AA318D
+WIF private key: 5JhK3KLBUcGgDZNTo2kmeEG6Xge1CFsdSWGncDC9on8uDvQPYUE
+public key: 04EF9D9B611648B359AF44A2D655D727C6EB5582EDD4651C918DDC14ED1685C0E2BC2574FBB94B46CF05C337C8520D7B29392255CDE0D97DE682538229A3DA6A90
+uncompressed address: 1DfZ8MeYCD9UaipUra5S24QuESwYZwbMiW
+Almost there!
+
+hex private key: 462CC279D4FB31C4E55C34D819753BDC458D9CE90AED5327BD7322A7761EEFB0
+WIF private key: 5JMC6y1yxpWTUHjRDh13YmZa34rZFkAcyXfsNNMy9rBbsdBbUPM
+public key: 0477B70308EFC410B454FCEF4F825C06F08D8085FA20104B00CD2033C551B2C23C9C12F3538E599633BC320A5BD223F02DAE1680DC6DCA3F07CE090958963FB560
+uncompressed address: 1gjmowKinvkHvYTAq8JK6nyGpxcc51HXi
+Almost there!
+
+hex private key: B1514268147365FB080ED39FB68A9B114A1FDDD7958130B505D36EBC794BDC08
+WIF private key: 5KANv94Yk6uAysVu5jv2WZGo39C7smYJKGTcEKv9Sed6Gyi8rMK
+public key: 04E910B264F7212CD325E43A8219761E252496B752C3C1202FC8A758B63912705514E67D9C5C64757879E71F8832A0464001D8106D4CF7D4C9EFCE1786D5099580
+uncompressed address: 16S8NyYNcBCu6VLuYHNxAgyo4wh5npQb17
+Almost there!
+
+hex private key: F347AAFD61790EFD72488977F5F3A0FED1C35E769AADD386C5C79E03D3DE06E8
+WIF private key: 5KfRqkgeUNA3xwrRPc7haeq7vebs7aGFnRj75yJgK4ekq3b3npF
+public key: 043BA74B47AAC54F117413FF8213BC90BAA3665FDA56599554F97A938D646483BDF1129A6C601E00D9AE6F82BC9D91089311DB313990C4368B5A2978E837F26C22
+uncompressed address: 18WU9dddMT2F4iLDCfxFJPWDepfGrbAHP3
+Almost there!
+
