+++ worker/database/README.md
@@ -1,0 +1,13 @@
+# Database FAQ
+
+This database is a collection of text files containing all funded Bitcoin addresses downloaded from http://addresses.loyce.club/
+
+The name of the text file is the date when the database was last updated in month_day_year format. The file will be updated every 3-6 months.
+
+### How Many Addresses Does The Database Have?
+
+The database currently holds `43,212,954 Bitcoin addresses`. This is the total number of Bitcoin addresses with a balance that exist in the blockchain at the time of uploading.
+
+### Why Are There So Many Files?
+
+There are multiple files because GitHub limits file uploads to 50 MB. The single file is too large, so it was split into multiple files each under 50 MB in order to be uploaded to GitHub.
